Refactor prune_nodes out of h_spark
This functionlity will be shared with pandas and polars, putting it
central.
